+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file
 - The selected password file organisation (in General settings) is now honoured when new entries are manually created
 - During autofill, login form fields are now pre-filled with the newly created credentials
 - In case of file (work/example.org/john@doe.org) and dir-based (work/example.org/john@dow.org/password.gpg) password file organisation, the subfolder structure is now preserved when modifying the suggested destination folder on the "New password" form
+- Improved username lookup for autofill, now based on origin (web/app form) and independent from current password file organisation setting
 
 ## [2.0.1] - 2026-08-07